Don't panic on jailing
The staking keeper receives a copy of the bank keeper, before it has been configured to burn tokens correctly. This causes a panic if it ever tries to burn tokens - e.g. if it is jailing a validator. This patch passes the bank keeper to the staking keeper by reference instead.